[C++]: manually use fastMod3 for indexBy functions. Add DISABLE_BOUNDS_CHECKS define for turning off boundsCheck in AtomicBuffer. Added DISABLE_BOUNDS_CHECKS to Release builds.

+++ aeron-samples/src/main/cpp/StreamingPublisher.cpp
@@ -198,7 +198,7 @@ int main(int argc, char **argv)
             }
 
             std::cout << "Done streaming. Back pressure ratio ";
-            std::cout << ((double)backPressureCount / (double)settings.numberOfMessages) << std::endl;
+            std::cout << ((double)backPressureCount / settings.numberOfMessages) << std::endl;
 
             if (running && settings.lingerTimeoutMs > 0)
             {
